refactor: use UtxoConfirmations in ShowUtxos component
