+++ src/components/utxo/Confirmations.module.css
@@ -1,16 +1,10 @@
 .confirmations {
   width: 2rem;
   display: flex;
-  gap: 0.1rem;
   flex-direction: column;
   justify-content: center;
   align-items: center;
   padding: 0.1rem 0.2rem;
   border-radius: 0.2rem;
-  color: var(--bs-success);
-  font-size: 0.6rem;
-}
-
-.confirmations-0 {
-  color: var(--bs-info);
+  font-size: 0.75rem;
 }
